refactor: 경험 상세 UI 섹션 컴포넌트 분리
